Create ResultsTextViewManager
